% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense_covs.R
\name{dense_covs}
\alias{dense_covs}
\title{Dense Covariance Difference Test : covid [1,4]}
\usage{
dense_covs(p, n, mreplicate, covid, rhos = c(seq(0.01, 2, length.out =
  20), seq(2.2, 5, length.out = 10)))
}
\description{
estA = Schott k-sample    (2007)
 estB = Li and Chen        (2012)
 estC = Cai, Liu, and Xia  (2013)
 estX = mxPBF
}
\details{
@examples
 ## run
 out = dense_covs(p=5, n=25, mreplicate=100, covid=1)
}
