% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense_mean.R
\name{dense_mean}
\alias{dense_mean}
\title{Dense Mean Difference Test : covid [1,8]}
\usage{
dense_mean(p, n, mreplicate, covid, rhos = c(seq(0.01, 2, length.out =
  20), seq(2.2, 5, length.out = 10)))
}
\description{
estA = Bai and Saranadasa (1996)
 estB = Srivastava and Du  (2008)
 estC = Cai, Liu, and Xia  (2014)
 estX = mxPBF
}
\examples{
# run
out = dense_mean(p=5, n=25, mreplicate=100, covid=1)

}
