% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_mean.R
\name{sp_mean}
\alias{sp_mean}
\title{Sparse Mean Difference Test : covid [1,8]}
\usage{
sp_mean(p, n, mreplicate, covid, ndev = 1, rhos = c(seq(0.01, 2,
  length.out = 20), seq(2.2, 5, length.out = 10)))
}
\description{
estA = Bai and Saranadasa (1996)
 estB = Srivastava and Du  (2008)
 estC = Cai, Liu, and Xia  (2014)
 estX = mxPBF
}
\details{
@examples
 ## run
 out = sp_mean(5, 25, 100, 1)
}
